chore: split manifest errors from definition
